functionality at 99%

+++ libft/ft_strcpy.c
@@ -6,13 +6,15 @@
 /*   By: WSeegers <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/04/16 23:58:13 by WSeegers          #+#    #+#             */
-/*   Updated: 2018/04/17 00:07:42 by WSeegers         ###   ########.fr       */
+/*   Updated: 2018/04/23 15:12:08 by WSeegers         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 char	*ft_strcpy(char *dest, const char *src)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (src[i])
